Split signals and events
