+++ crates/lib/src/lib.rs
@@ -113,7 +113,7 @@ pub mod handler;
 mod watchexec;
 
 // compatibility
-#[deprecated(note = "use the `watchexec_events` crate directly instead")]
+#[deprecated(note = "use the `watchexec-events` crate directly instead", since = "2.1.0")]
 pub use watchexec_events as event;
 
 #[doc(inline)]
